AI Engine refactoring and FrameBuffer Implement
