+++ server/dpgs-server/ai_engine/config/map.json
@@ -1,0 +1,21 @@
+{
+  "parking_lot_id": "PL123",
+  "total_slots": 3,
+  "slots": [
+    {
+      "slot_id": 1,
+      "state": "occupied",
+      "poly": [[56, 683], [593, 668], [672, 492], [267, 502]]
+    },
+    {
+      "slot_id": 2,
+      "state": "vacant",
+      "poly": [[605, 668], [1136, 649], [1074, 480], [679, 494]]
+    },
+    {
+      "slot_id": 3,
+      "state": "vacant",
+      "poly": [[1149, 648], [1688, 635], [1489, 468], [1086, 476]]
+    }
+  ]
+}
