+++ server/dpgs-server/ai_engine/tb/vpe_dummy.cpp
@@ -1,0 +1,46 @@
+#include "vpe_dummy.h"
+#include <chrono>
+#include <iostream>
+
+
+const int target_fps = 30;
+const int delay_ms = 1000 / target_fps;
+
+
+VpeDummy::VpeDummy(const std::string& _path, FrameBuffer& _fb)
+    : path(_path), fb(_fb) {
+}
+
+
+void VpeDummy::start() {
+    run();
+}
+
+
+void VpeDummy::run() {
+    cv::VideoCapture test_video(path);
+    if (!test_video.isOpened()) {
+        std::cerr << "[VPED] Error: Failed to open video: " << path << "\n";
+        return;
+    }
+
+    std::cout << "[VPED] Success: test_video is opened and start to push\n";
+
+    cv::Mat frame;
+    while (true) {
+        if (!test_video.read(frame)) {
+            test_video.set(cv::CAP_PROP_POS_FRAMES, 0);
+            continue;
+        }
+
+        // Debug Session
+//        size_t size = frame.total() * frame.elemSize();
+//        std::cout << "[DEBUG][VPED] push: Frame Size: " << size << ", total: " << frame.total() << ", rows: " << frame.rows << ", cols: " << frame.cols <<  "\n";
+        // ------
+
+        fb.push(frame);
+        std::this_thread::sleep_for(std::chrono::milliseconds(delay_ms));
+    }
+
+    test_video.release();
+}
